Rename file to correspond to class name
